move server actions to correct dir
